+++ pytalk/instance.py
@@ -13,11 +13,22 @@ import threading
 import time
 from typing import List, Optional, Union
 
-from ._utils import _do_after, _waitForCmd, _waitForEvent, percent_to_ref_volume, ref_volume_to_percent
-
-from .audio import AudioBlock, MuxedAudioBlock, _AcquireUserAudioBlock, _ReleaseUserAudioBlock
+from ._utils import (
+    _do_after,
+    _waitForCmd,
+    _waitForEvent,
+    percent_to_ref_volume,
+    ref_volume_to_percent,
+)
+from .audio import (
+    AudioBlock,
+    MuxedAudioBlock,
+    _AcquireUserAudioBlock,
+    _ReleaseUserAudioBlock,
+)
 from .channel import Channel as TeamTalkChannel
 from .channel import ChannelType
+from .codec import CodecType
 from .device import SoundDevice
 from .enums import TeamTalkServerInfo, UserStatusMode, UserType
 from .exceptions import PermissionError
@@ -296,6 +307,35 @@ class TeamTalkInstance(sdk.TeamTalk):
         if not success:
             _log.error(f"Failed to set input volume for instance {self.server_info.host}")
         return success
+
+    # Media File Streaming Functions
+    def start_streaming_media_file_to_channel(self, path: str, video_codec: Optional[sdk.VideoCodec] = None) -> bool:
+        """Starts streaming a media file to the channel.
+
+        If no video codec is specified, it defaults to WebM VP8.
+
+        Args:
+            path (str): The path to the media file.
+            video_codec (Optional[sdk.VideoCodec]): An optional video codec object.
+
+        Returns:
+            bool: True if the streaming started successfully, False otherwise.
+        """
+        if video_codec is None:
+            codec_to_use = sdk.VideoCodec()
+            codec_to_use.nCodec = CodecType.WEBM_VP8
+        else:
+            codec_to_use = video_codec
+
+        return self.super.startStreamingMediaFileToChannel(sdk.ttstr(path), ctypes.byref(codec_to_use))
+
+    def stop_streaming_media_file_to_channel(self) -> bool:
+        """Stops the current media file streaming to the channel.
+
+        Returns:
+            bool: True if the streaming stopped successfully, False otherwise.
+        """
+        return self.super.stopStreamingMediaFileToChannel()
 
     # permission stuff
     def has_permission(self, permission: Permission) -> bool: